[변경] 알림 디코딩 모델 추가 및 목 데이터 삭제

+++ SOOUM/SOOUM/Domain/Models/CommonNotificationInfo.swift
@@ -50,7 +50,7 @@ extension CommonNotificationInfo: Decodable {
     
     init(from decoder: any Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.notificationId = String(try container.decode(Int.self, forKey: .notificationId))
+        self.notificationId = String(try container.decode(Int64.self, forKey: .notificationId))
         self.notificationType = try container.decode(CommonNotificationInfo.NotificationType.self, forKey: .notificationType)
         self.createTime = try container.decode(Date.self, forKey: .createTime)
     }
